fix ios build
