+++ app/src/main/java/dev/dimension/flare/ui/screen/compose/ComposeScreen.kt
@@ -799,7 +799,7 @@ private fun ComposeScreen(
                                             ),
                                     ) {
                                         EmojiPicker(
-                                            data = emojis,
+                                            data = emojis.data,
                                             onEmojiSelected = state::selectEmoji,
                                             modifier =
                                                 Modifier
