+++ shared/src/commonMain/kotlin/dev/dimension/flare/ui/presenter/settings/AccountsPresenter.kt
@@ -4,7 +4,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
+import dev.dimension.flare.common.ImmutableListWrapper
 import dev.dimension.flare.common.collectAsState
+import dev.dimension.flare.common.toImmutableListWrapper
 import dev.dimension.flare.data.repository.AccountRepository
 import dev.dimension.flare.data.repository.accountServiceProvider
 import dev.dimension.flare.data.repository.activeAccountPresenter
@@ -18,7 +20,6 @@ import dev.dimension.flare.ui.model.flatMap
 import dev.dimension.flare.ui.model.map
 import dev.dimension.flare.ui.model.toUi
 import dev.dimension.flare.ui.presenter.PresenterBase
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
@@ -72,21 +73,3 @@ public abstract class AccountsState(
 
     public abstract fun removeAccount(accountKey: MicroBlogKey)
 }
-
-@Immutable
-public data class ImmutableListWrapper<T : Any>(
-    private val data: ImmutableList<T>,
-) {
-    val size: Int
-        get() = data.size
-
-    public operator fun get(index: Int): T = data[index]
-
-    public fun indexOf(element: T): Int = data.indexOf(element)
-
-    public fun contains(element: T): Boolean = data.contains(element)
-
-    public fun toImmutableList(): ImmutableList<T> = data
-}
-
-internal fun <T : Any> ImmutableList<T>.toImmutableListWrapper(): ImmutableListWrapper<T> = ImmutableListWrapper(this)
