+++ shared/src/commonMain/kotlin/dev/dimension/flare/ui/presenter/status/action/MisskeyReactionPresenter.kt
@@ -9,6 +9,7 @@ import dev.dimension.flare.data.repository.AccountRepository
 import dev.dimension.flare.data.repository.accountServiceProvider
 import dev.dimension.flare.model.AccountType
 import dev.dimension.flare.model.MicroBlogKey
+import dev.dimension.flare.ui.model.EmojiData
 import dev.dimension.flare.ui.model.UiEmoji
 import dev.dimension.flare.ui.model.UiState
 import dev.dimension.flare.ui.model.UiTimeline
@@ -18,8 +19,6 @@ import dev.dimension.flare.ui.model.onSuccess
 import dev.dimension.flare.ui.model.toUi
 import dev.dimension.flare.ui.presenter.PresenterBase
 import dev.dimension.flare.ui.presenter.status.StatusPresenter
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.ImmutableMap
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
@@ -41,11 +40,16 @@ public class MisskeyReactionPresenter(
                 service as MisskeyDataSource
             }
         val data =
-            service.flatMap {
-                remember(it) {
-                    it.emoji()
-                }.collectAsState().toUi()
-            }
+            service
+                .flatMap {
+                    remember(it) {
+                        it.emoji()
+                    }.collectAsState().toUi()
+                }.map {
+                    remember(it) {
+                        EmojiData(it)
+                    }
+                }
 
         val status =
             remember(statusKey, accountType) {
@@ -80,7 +84,7 @@ public class MisskeyReactionPresenter(
 
 @Immutable
 public interface MisskeyReactionState {
-    public val emojis: UiState<ImmutableMap<String, ImmutableList<UiEmoji>>>
+    public val emojis: UiState<EmojiData>
 
     public fun select(emoji: UiEmoji)
 }
